adapted sdk generation to unix environment

+++ scripts/generate-nestjs-sdk.js
@@ -19,35 +19,50 @@ function getCurrentDirectory() {
   return "$(pwd)";
 }
 
-// NOTE: First do some cleanup before starting the generation
-console.log("Cleanup old files...");
-execSync(
-  "rm -rf node_modules/@scicatproject/scicat-sdk-ts && rm -rf @scicatproject/scicat-sdk-ts",
-  { encoding: "utf-8" },
-);
+if (isWindows()) {
 
-console.log("Generating the new sdk...");
-const generationOutput = execSync(
-  `docker run --rm --add-host host.docker.internal:host-gateway -v "${getCurrentDirectory()}:/local" openapitools/openapi-generator-cli:v7.9.0 generate -i http://host.docker.internal:3000/explorer-json -g typescript-angular -o local/@scicatproject/scicat-sdk-ts --additional-properties=ngVersion=16.2.12,npmName=@scicatproject/scicat-sdk-ts,supportsES6=true,npmVersion=10.8.2,withInterfaces=true`,
-  { encoding: "utf-8" },
-);
-console.log(generationOutput);
+  // NOTE: First do some cleanup before starting the generation
+  console.log("Cleanup old files...");
+  execSync(
+    "rm -rf node_modules/@scicatproject/scicat-sdk-ts && rm -rf @scicatproject/scicat-sdk-ts",
+    { encoding: "utf-8" },
+  );
 
-console.log("Installing dependencies and building the sdk...");
-const installBuildOutput = execSync(
-  "cd @scicatproject/scicat-sdk-ts && npm install && npm run build",
-  { encoding: "utf-8" },
-);
-console.log(installBuildOutput);
+  console.log("Generating the new sdk...");
+  const generationOutput = execSync(
+    `docker run --rm --add-host host.docker.internal:host-gateway -v "${getCurrentDirectory()}:/local" openapitools/openapi-generator-cli:v7.9.0 generate -i http://host.docker.internal:3000/explorer-json -g typescript-angular -o local/@scicatproject/scicat-sdk-ts --additional-properties=ngVersion=16.2.12,npmName=@scicatproject/scicat-sdk-ts,supportsES6=true,npmVersion=10.8.2,withInterfaces=true`,
+    { encoding: "utf-8" },
+  );
+  console.log(generationOutput);
 
-console.log("Copying the build files in node_modules...");
-const copyToNodeModulesOutput = execSync(
-  "cp -r @scicatproject/scicat-sdk-ts/dist node_modules/@scicatproject/scicat-sdk-ts",
-  { encoding: "utf-8" },
-);
-console.log(copyToNodeModulesOutput);
+  console.log("Installing dependencies and building the sdk...");
+  const installBuildOutput = execSync(
+    "cd @scicatproject/scicat-sdk-ts && npm install && npm run build",
+    { encoding: "utf-8" },
+  );  
+  console.log(installBuildOutput);
 
-console.log("Final cleanup...");
-execSync("rm -rf @scicatproject", {
-  encoding: "utf-8",
-});
+  console.log("Copying the build files in node_modules...");
+  const copyToNodeModulesOutput = execSync(
+    "cp -r @scicatproject/scicat-sdk-ts/dist node_modules/@scicatproject/scicat-sdk-ts",
+    { encoding: "utf-8" },
+  );
+  console.log(copyToNodeModulesOutput);
+
+  console.log("Final cleanup...");
+  execSync("rm -rf @scicatproject", {
+    encoding: "utf-8",
+  });
+
+  console.log("Local SDK generation completed");
+
+} else {
+  console.log("Your environment is a linux/unix");
+  console.log("Please run the following command on your terminal:");
+  console.log("> sudo -E ./scripts/generate-nestjs-sdk.bash");
+  console.log("");
+  console.log("IMPORTANT: the script runs under sudo. You will be asked your password.");
+  console.log("");
+
+}
+
